requirements.txt
Django==3.0.5
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django==3.0.5
django-nose==1.4.6
gunicorn==20.0.4
urllib3==1.25.9
whitenoise==5.0.1
autopep8==1.5.5

